+++ src/main/java/br/comexport/avaliacao/services/RoleService.java
@@ -1,9 +1,12 @@
 package br.comexport.avaliacao.services;
 
 import br.comexport.avaliacao.entities.RoleEntity;
+import br.comexport.avaliacao.entities.UserEntity;
 import br.comexport.avaliacao.exception.ResourceNotFoundException;
 import br.comexport.avaliacao.parameters.RoleParameter;
 import br.comexport.avaliacao.repositories.RoleRepository;
+import br.comexport.avaliacao.repositories.UserRepository;
+import br.comexport.avaliacao.util.Util;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,7 +19,17 @@ public class RoleService {
     @Autowired
     RoleRepository roleRepository;
 
+    @Autowired
+    UserRepository userRepository;
+
+    @Autowired
+    Util util;
+
     public ResponseEntity<Object> applySave(RoleParameter roleParameter)  {
+
+        ResponseEntity<Object> isvalid = validRole(roleParameter);
+        if(isvalid != null)
+            return isvalid;
 
         RoleEntity roleEntity = new RoleEntity();
         roleEntity.setDescription(roleParameter.getDescription());
@@ -36,24 +49,48 @@ public class RoleService {
     }
 
     public ResponseEntity<?> deleteRole(Long id){
+
         RoleEntity role = roleRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Role", "id", id));
 
-        roleRepository.delete(role);
+        List<UserEntity> userEntities = userRepository.selectUserRole(id);
 
-        return ResponseEntity.ok().build();
+        if (userEntities.size() == 0) {
+
+            roleRepository.delete(role);
+
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(util.exceptError("Dados inválidos","Role não pode ser excluída antes de excluir o usuário!"));
+        }
+
     }
 
     public ResponseEntity<Object> updateRole(Long id, RoleParameter roleParameter){
         RoleEntity role = roleRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Role", "id", id));
+
+        ResponseEntity<Object> isvalid = validRole(roleParameter);
+        if(isvalid != null)
+            return isvalid;
+
         RoleEntity roleEntity = new RoleEntity();
+        roleEntity.setId(id);
         roleEntity.setDescription(roleParameter.getDescription());
-        roleEntity.setEnabled(true);
+        roleEntity.setEnabled(roleParameter.isEnabled());
 
         roleRepository.save(roleEntity);
         return ResponseEntity.status(HttpStatus.OK)
                 .body(roleEntity);
+    }
+
+    public ResponseEntity<Object> validRole(RoleParameter roleParameter){
+        if ( roleParameter.getDescription().isEmpty() ) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(util.exceptError("Dados inválidos","Descrição não foi informada!"));
+        }
+        return null;
     }
 
 }
